добавляет, настраивает webpack, webpack-dev-server. Изменяет настройки eslint, package,json, gitignore. Разбивает проект на прото-модули
