+++ src/components/main/films/button.js
@@ -1,0 +1,5 @@
+const createFilmsButtonTemplate = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
+
+export {createFilmsButtonTemplate};
